+++ packages/eyes-storybook/src/yargsOptions.js
@@ -49,6 +49,19 @@ module.exports = {
     number: true,
   },
 
+  include: {
+    description:
+      'A string that represents a story title or a regex (starts and ends with /) to match stories title',
+    requiresArg: false,
+    string: true,
+    coerce: function(arg) {
+      if (arg.startsWith('/') && arg.endsWith('/')) {
+        // create a regex and remove slashes from the start and end of the input
+        return new RegExp(arg.substring(1, arg.length - 1));
+      } else return arg;
+    },
+  },
+
   // general
   exitcode: {
     alias: 'e',