+++ packages/eyes-storybook/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- support passing a string or regex to the include configuration
+
 
 ## 3.18.0 - 2021/2/26